sync: split CmuxSyncStore + tests under the 500-line file guard
Move the low-level SQLite statement helpers (BindValue, exec/bind/
transaction, user_version + cursor/tombstone writers) into
CmuxSyncStore+SQLite.swift, and split the test suites into
SyncFrameAndProtocolTests.swift, so both the store (490) and each test
file (<500) stay under the swift-file-length-budget threshold without a
budget bump. Pure code-motion; behavior unchanged.
